+++ app/src/main/java/com/example/tasky/common/presentation/Header.kt
@@ -1,0 +1,31 @@
+package com.example.tasky.common.presentation
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
+
+
+@Composable
+fun Header(title: String) {
+    Text(
+        text = title,
+        style = TextStyle(
+            fontSize = 28.sp,
+            fontWeight = FontWeight.Bold,
+            color = Color.White
+        )
+    )
+}
+
+@Composable
+@Preview
+fun ViewHeader() {
+    Column {
+        Header(title = "HEADLINE")
+    }
+}
