+++ app/src/main/java/com/example/tasky/feature_login/presentation/LoginScreen.kt
@@ -7,65 +7,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import com.example.tasky.common.presentation.Header
+import com.example.tasky.common.presentation.SimpleButton
+import com.example.tasky.common.presentation.TextBox
 
-@Composable
-fun Header(title: String) {
-    Text(
-        text = title,
-        style = androidx.compose.ui.text.TextStyle(
-            fontSize = 28.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color.White
-        )
-    )
-}
-
-@Composable
-fun TextField() {
-    var text by remember { mutableStateOf("") }
-
-    TextField(
-        value = text,
-        onValueChange = { text = it },
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        label = { Text("Enter text") }
-    )
-}
-
-@Composable
-fun Button(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    buttonName: String
-) {
-    Button(
-        onClick = onClick,
-        modifier = modifier.padding(16.dp),
-        enabled = enabled,
-        colors = ButtonDefaults.buttonColors(),
-    ) {
-        Text(buttonName)
-    }
-}
 
 @Composable
 @Preview
@@ -73,19 +24,24 @@ fun LoginScreenContent() {
     Column(
         modifier = Modifier
             .fillMaxSize()
-            .padding(top = 70.dp)
-            .background(Color.Black),
+            .background(Color.Black)
+            .padding(top = 70.dp),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
         Header(title = "Welcome Back!")
+        Column {
+
+        }
         Spacer(modifier = Modifier.height(16.dp))
         // build textfield
-        TextField()
-        Spacer(modifier = Modifier.height(16.dp))
-        // build textfield
-        TextField()
-        Spacer(modifier = Modifier.height(16.dp))
+        TextBox(hintText = "Email")
+        TextBox(hintText = "Password")
         // login button
-        Button(modifier = Modifier.fillMaxWidth(), onClick = { /*TODO*/ }, buttonName = "Test")
+        Spacer(modifier = Modifier.height(16.dp))
+        SimpleButton(
+            modifier = Modifier.fillMaxWidth(),
+            onClick = { /*TODO*/ },
+            buttonName = "Test"
+        )
     }
 }
